# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/courses": {
      "post": {
        "operationId": "CourseController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created successfully."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new course (Admin only)",
        "tags": [
          "Courses"
        ]
      },
      "get": {
        "operationId": "CourseController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by title",
            "schema": {
              "example": "typescript",
              "type": "string"
            }
          },
          {
            "name": "difficultyLevel",
            "required": false,
            "in": "query",
            "description": "Filter by difficulty level",
            "schema": {
              "type": "string",
              "enum": [
                "beginner",
                "intermediate",
                "advanced"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of courses."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all courses (paginated)",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/{id}": {
      "get": {
        "operationId": "CourseController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course details."
          },
          "404": {
            "description": "Course not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a course by ID",
        "tags": [
          "Courses"
        ]
      },
      "put": {
        "operationId": "CourseController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course updated."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Course not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a course (Admin only)",
        "tags": [
          "Courses"
        ]
      },
      "delete": {
        "operationId": "CourseController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course deleted."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Course not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a course (Admin only)",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/{id}/tutors": {
      "get": {
        "operationId": "CourseController_getTutors",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tutors."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all tutors for a course",
        "tags": [
          "Courses"
        ]
      },
      "post": {
        "operationId": "CourseController_assignTutor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTutorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tutor assigned."
          },
          "404": {
            "description": "Course or tutor not found."
          },
          "409": {
            "description": "Tutor already assigned."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign a tutor to a course",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/{id}/tutors/bulk": {
      "post": {
        "operationId": "CourseController_bulkAssignTutors",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkAssignTutorsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tutors assigned."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bulk assign tutors to a course",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/{id}/tutors/{tutorId}": {
      "delete": {
        "operationId": "CourseController_removeTutor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tutorId",
            "required": true,
            "in": "path",
            "description": "Tutor ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tutor removed."
          },
          "404": {
            "description": "Assignment not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a tutor from a course",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/{id}/learners": {
      "get": {
        "operationId": "CourseController_getLearners",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cohortId",
            "required": false,
            "in": "query",
            "description": "Filter by cohort ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of learners grouped by cohort."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all learners for a course (optionally by cohort)",
        "tags": [
          "Courses"
        ]
      },
      "post": {
        "operationId": "CourseController_enrollLearner",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollLearnerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Learner enrolled."
          },
          "404": {
            "description": "Course, cohort, or learner not found."
          },
          "409": {
            "description": "Learner already enrolled."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enroll a learner in a course cohort",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/{id}/learners/bulk": {
      "post": {
        "operationId": "CourseController_bulkEnrollLearners",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkEnrollLearnersDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Learners enrolled."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bulk enroll learners in a course cohort",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/{id}/learners/{cohortId}/{learnerId}": {
      "delete": {
        "operationId": "CourseController_removeLearner",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cohortId",
            "required": true,
            "in": "path",
            "description": "Cohort ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "learnerId",
            "required": true,
            "in": "path",
            "description": "Learner ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Learner removed."
          },
          "404": {
            "description": "Enrollment not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a learner from a course cohort",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/{id}/contents": {
      "get": {
        "operationId": "CourseController_getContents",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of course contents."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all contents for a course (tree structure)",
        "tags": [
          "Courses"
        ]
      },
      "post": {
        "operationId": "CourseController_createContent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseContentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Content created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create course content (Admin/Assigned Tutor only)",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/contents/{contentId}": {
      "put": {
        "operationId": "CourseController_updateContent",
        "parameters": [
          {
            "name": "contentId",
            "required": true,
            "in": "path",
            "description": "Course Content ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseContentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Content updated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update course content (Admin/Assigned Tutor only)",
        "tags": [
          "Courses"
        ]
      },
      "delete": {
        "operationId": "CourseController_removeContent",
        "parameters": [
          {
            "name": "contentId",
            "required": true,
            "in": "path",
            "description": "Course Content ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Content deleted."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete course content (Admin/Assigned Tutor only)",
        "tags": [
          "Courses"
        ]
      }
    },
...
  },
  "info": {
    "title": "LMS API",
    "description": "## Learning Management System REST API\n\nThis API powers the LMS platform, supporting:\n- **Authentication** – signup, login, password reset\n- **Users** – CRUD for all user roles (SuperAdmin, Admin, Tutor, Learner)\n- **Courses** – course management, tutor assignment, learner enrollment, content tree\n- **Assessments** – quizzes, code challenges, group Kahoot-style sessions\n- **AI** – audio transcription, quiz/flashcard generation, performance analysis\n\n### Authentication\nAll protected endpoints require a **Bearer JWT token**.  \nObtain a token via `POST /auth/login` and pass it as:\n```\nAuthorization: Bearer <token>\n```\n\n### Roles\n| Role | Description |\n|------|-------------|\n| `superadmin` | Full access |\n| `admin` | Manage courses, users, assessments |\n| `tutor` | Manage assigned course content |\n| `learner` | View content, submit assessments |\n",
    "version": "1.0",
    "contact": {
      "name": "LMS Team",
      "url": "",
      "email": "support@lms.example.com"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter your JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Introduction to TypeScript",
            "description": "Course title"
          },
          "description": {
            "type": "string",
            "example": "Learn TypeScript from scratch",
            "description": "Course description"
          },
          "difficultyLevel": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "example": "beginner",
            "description": "Difficulty level"
          }
        },
        "required": [
          "title",
          "difficultyLevel"
        ]
      },
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Advanced TypeScript",
            "description": "Course title"
          },
          "description": {
            "type": "string",
            "example": "Deep dive into TypeScript",
            "description": "Course description"
          },
          "difficultyLevel": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "description": "Difficulty level"
          }
        }
      },
      "AssignTutorDto": {
        "type": "object",
        "properties": {
          "tutorId": {
            "type": "string",
            "example": "uuid-of-tutor",
            "description": "ID of the tutor to assign"
          }
        },
        "required": [
          "tutorId"
        ]
      },
      "BulkAssignTutorsDto": {
        "type": "object",
        "properties": {
          "tutorIds": {
            "example": [
              "uuid-1",
              "uuid-2"
            ],
            "description": "Array of tutor IDs to assign",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "tutorIds"
        ]
      },
      "EnrollLearnerDto": {
        "type": "object",
        "properties": {
          "learnerId": {
            "type": "string",
            "example": "uuid-of-learner",
            "description": "ID of the learner to enroll"
          },
          "cohortId": {
            "type": "string",
            "example": "uuid-of-cohort",
            "description": "ID of the cohort to enroll into"
          }
        },
        "required": [
          "learnerId",
          "cohortId"
        ]
      },
      "BulkEnrollLearnersDto": {
        "type": "object",
        "properties": {
          "learnerIds": {
            "example": [
              "uuid-1",
              "uuid-2"
            ],
            "description": "Array of learner IDs to enroll",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "cohortId": {
            "type": "string",
            "example": "uuid-of-cohort",
            "description": "ID of the cohort to enroll into"
          }
        },
        "required": [
          "learnerIds",
          "cohortId"
        ]
      },
      "CreateCourseContentDto": {
        "type": "object",
        "properties": {
          "topic": {
            "type": "string",
            "example": "Module 1: Getting Started",
            "description": "Content topic"
          },
          "contentType": {
            "type": "string",
            "enum": [
              "section",
              "lesson",
              "assessment"
            ],
            "description": "Type of content"
          },
          "sequenceOrder": {
            "type": "number",
            "example": 1,
            "description": "Order of priority"
          },
          "parentId": {
            "type": "string",
            "example": "uuid-of-parent",
            "description": "Parent content ID for nested sections"
          }
        },
        "required": [
          "topic",
          "contentType"
        ]
      },
      "UpdateCourseContentDto": {
        "type": "object",
        "properties": {
          "topic": {
            "type": "string",
            "example": "Module 1: Updated Topic",
            "description": "Content topic"
          },
          "contentType": {
            "type": "string",
            "enum": [
              "section",
              "lesson",
              "assessment"
            ],
            "description": "Type of content"
          },
          "sequenceOrder": {
            "type": "number",
            "example": 2,
            "description": "Order of priority"
          },
          "parentId": {
            "type": "string",
            "example": "uuid-of-parent",
            "description": "Parent content ID"
          }
        }
      },
...
    }
  }
}